output/dnssim: refactor - move tcp to separate file
